records: merge orcid receivers
