+++ inspirehep/modules/orcid/__init__.py
@@ -27,6 +27,4 @@ from __future__ import (
 
 from .ext import INSPIREOrcid
 
-from .receivers import *
-
 __all__ = ('INSPIREOrcid',)
